backend/requirements.txt
aniso8601==7.0.0
asgiref==1.1.2
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
awscli==1.16.132
botocore==1.12.122
certifi==2019.6.16
cffi==1.12.3
channels==1.1.8
channels-api==0.4.1
colorama==0.3.9
constantly==15.1.0
cryptography==2.7
daphne==1.4.2
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
django-graphql-jwt==0.2.2
django-model-utils==3.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
graphene-django-subscriptions==0.0.6
graphene==2.1.8
graphene-django==2.5.0
graphene-django-extras==0.4.6
graphql-core==2.2.1
graphql-relay==2.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
iso8601==0.1.12
jmespath==0.9.4
jsonfield==2.0.2
mysqlclient==1.4.4
promise==2.2.1
psycopg2-binary==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==3.13
rsa==3.4.2
Rx==1.6.1
s3transfer==0.2.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
zope.interface==4.6.0
